factory: split factory_review.py to satisfy 400-line CI check
The rename commit (60693934) touched factory_review.py, which triggered the
repo-wide file-size check on a module that was already 526 lines at
origin/main. Extract the pure review-spec helpers (detect_actionable_findings,
trim_detail, pick_secondary_lens, required_reviews) and their model/threshold
constants into a new factory_review_specs.py module. factory_review.py
re-exports those names so run_factory and the test suite see no API change.

After the split:
- factory_review.py: 398 lines (was 526)
- factory_review_specs.py: 151 lines (new)
- All 122 factory repair tests pass
- CI file-size check passes locally
